Move bind inputs func out to seperate file
